renaming sk service.
